+++ tests/ByteDev.Testing.IntTests/DummyKeyVaultSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace ByteDev.Testing.IntTests
+{
+    public class DummyKeyVaultSettings
+    {
+        public string SdkUserIdGrab { get; set; }
+    }
+}